+++ app/src/main/java/com/example/onenthapp/MainActivity.kt
@@ -7,13 +7,14 @@ import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
-import androidx.core.os.bundleOf
 import androidx.core.view.isVisible
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.NavigationUI
 import androidx.navigation.ui.setupWithNavController
 import com.example.onenthapp.feature.chat.ChatActivity
 import com.example.onenthapp.databinding.ActivityMainBinding
+import com.example.onenthapp.feature.board.CreateLifePostActivity
+import com.example.onenthapp.feature.board.TipFragment
 import com.example.onenthapp.model.HomeTabType
 import com.example.onenthapp.model.SharedViewModel
