+++ app/src/main/java/com/example/onenthapp/MyPageReviewsActivity.kt
@@ -13,10 +13,8 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.example.onenthapp.data.MyReview
 import com.example.onenthapp.databinding.ActivityMypageReviewsBinding
+import com.example.onenthapp.feature.mypage.ReviewEditActivity
 import com.example.onenthapp.util.TokenManager
-import kotlinx.coroutines.async
-import kotlinx.coroutines.awaitAll
-import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
 
 class MyPageReviewsActivity : AppCompatActivity() {
