+++ app/src/main/java/com/example/onenthapp/MyReviewAdapter.kt
@@ -5,12 +5,12 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
-import android.widget.LinearLayout
 import android.widget.RatingBar
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.example.onenthapp.data.MyReview
+import com.example.onenthapp.feature.mypage.ReviewEditActivity
 
 class MyReviewAdapter(
     private var reviewList: List<MyReview>  // ← updateList 위해 var
