+++ app/src/main/java/com/example/onenthapp/SplashActivity.kt
@@ -10,6 +10,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import com.example.onenthapp.data.login.AuthRepository
 import com.example.onenthapp.data.login.ReissueResponse
+import com.example.onenthapp.feature.login.FindAccountActivity
 import com.example.onenthapp.feature.login.LoginActivity
 import com.example.onenthapp.feature.signup.SignupActivity
 import com.example.onenthapp.feature.signup.SignupActivity2
