+++ app/src/main/java/com/example/onenthapp/feature/board/DiscountTipsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.onenthapp
+package com.example.onenthapp.feature.board
 
 import android.content.Intent
 import android.graphics.Color
@@ -10,6 +10,7 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
+import com.example.onenthapp.R
 import com.example.onenthapp.data.map.*
 import com.example.onenthapp.databinding.FragmentTipsDiscounttipsBinding
 import com.example.onenthapp.databinding.ItemMarkerDetailBinding
