+++ app/src/main/java/com/example/onenthapp/feature/board/LifeTipsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.onenthapp
+package com.example.onenthapp.feature.board
 
 import android.content.Intent
 import android.os.Bundle
@@ -10,6 +10,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.onenthapp.RetrofitInstance
 import com.example.onenthapp.databinding.FragmentTipsLifetipsBinding
 import com.example.onenthapp.data.post.TipItem
 import com.example.onenthapp.util.TokenManager
