+++ app/src/main/java/com/example/onenthapp/feature/mypost/MyPostAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.onenthapp
+package com.example.onenthapp.feature.mypost
 
 import android.view.LayoutInflater
 import android.view.View
@@ -8,6 +8,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import com.example.onenthapp.R
 import com.example.onenthapp.data.post.MyPostItem
 
 // MyPostAdapter.kt
