+++ app/src/main/java/com/example/onenthapp/feature/mypost/MyPostPagerAdapter.kt
@@ -1,9 +1,10 @@
-package com.example.onenthapp
+package com.example.onenthapp.feature.mypost
 
 
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.example.onenthapp.feature.scrap.ScrapPostFragment
 
 class MyPostPagerAdapter(activity: AppCompatActivity) : FragmentStateAdapter(activity) {
     override fun getItemCount() = 2
