+++ app/src/main/java/com/example/onenthapp/feature/scrap/ScrapPagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.onenthapp
+package com.example.onenthapp.feature.scrap
 
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
