+++ app/src/main/java/com/example/onenthapp/feature/scrap/ScrapProductFragment.kt
@@ -1,4 +1,4 @@
-package com.example.onenthapp
+package com.example.onenthapp.feature.scrap
 
 import android.os.Bundle
 import android.view.View
@@ -7,6 +7,9 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.onenthapp.R
+import com.example.onenthapp.RetrofitInstance
+import com.example.onenthapp.ScrapProductAdapter
 import com.example.onenthapp.data.MyPostProductItem
 import com.example.onenthapp.util.TokenManager
 import kotlinx.coroutines.launch
