+++ app/src/main/res/layout/activity_edit_posts.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@color/main_white"
     android:orientation="vertical"
-    tools:context=".LifeTipsSearchActivity">
+    tools:context=".feature.board.LifeTipsSearchActivity">
 
     <!-- 툴바 -->
     <androidx.appcompat.widget.Toolbar
