+++ app/src/main/res/layout/activity_life_details.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/main_white"
-    tools:context=".LifeTipsSearchActivity">
+    tools:context=".feature.board.LifeTipsSearchActivity">
 
     <!-- 툴바 -->
     <androidx.appcompat.widget.Toolbar
